refactor(core,application): include hygiene + host/device separation
Clean up the core/ and application/ include graph; no logic changes.

- Layering: core/ no longer includes application/. Relocate the low-level SDMA
  device handle + DMA-barrier headers down into core (anvil_device.hpp,
  sdma_pkt_struct.h, udma_barrier.h).
- Dead code: remove the unused C++ Torch bootstrap path (also drops the libtorch
  link from libmori_application) and the dead host-side RDMA primitives
  (host_primitives.hpp / mlx5_host_primitives.hpp).
- Host/device separation: keep host-only deps out of device TUs — drop
  <hsakmt/hsakmt.h> (host KFD driver API) and the host ibverbs-DV bnxt_re_dv.h
  out of the device headers; split providers/utils.h so each provider's CQE
  decoder lives in its own provider header; drop the hsakmt/anvil drag from
  application_device_types.hpp; make primitives.hpp a pure ibverbs-handle header.
- IWYU: add direct includes for transitively-relied-on symbols, remove unused
  includes across core + application, add a device-only rdma_device.hpp aggregator.

Validated on BNXT: full build (all modules) clean; shmem P2P/IBGDA, ccl SDMA
allgather/all2all, ops intranode all pass.
