+++ examples/application/ibverbs_test.cpp
@@ -23,7 +23,7 @@
 #include <mpi.h>
 
 #include "mori/application/application.hpp"
-#include "mori/application/utils/udma_barrier.h"
+#include "mori/core/utils/udma_barrier.h"
 #include "mori/core/core.hpp"
 
 using namespace mori;
