+++ examples/benchmarks/atomic_perf.cpp
@@ -24,7 +24,7 @@
 
 #include "args_parser.hpp"
 #include "mori/application/application.hpp"
-#include "mori/application/utils/udma_barrier.h"
+#include "mori/core/utils/udma_barrier.h"
 #include "mori/core/core.hpp"
 
 using namespace mori;
