+++ examples/dist_rdma_ops/dist_write.cpp
@@ -26,7 +26,7 @@
 #include "args_parser.hpp"
 #include "mori/application/application.hpp"
 #include "mori/application/topology/topology.hpp"
-#include "mori/application/utils/udma_barrier.h"
+#include "mori/core/utils/udma_barrier.h"
 #include "mori/core/core.hpp"
 
 using namespace mori;
