+++ examples/local_rdma_ops/write_gpu.cpp
@@ -22,7 +22,7 @@
 #include <hip/hip_runtime.h>
 
 #include "mori/application/application.hpp"
-#include "mori/application/utils/udma_barrier.h"
+#include "mori/core/utils/udma_barrier.h"
 #include "mori/core/core.hpp"
 
 using namespace mori;
