+++ include/mori/application/bootstrap/bootstrap.hpp
@@ -25,7 +25,4 @@
 #ifdef MORI_WITH_MPI
 #include "mori/application/bootstrap/mpi_bootstrap.hpp"
 #endif
-#ifdef MORI_HAS_TORCH
-#include "mori/application/bootstrap/torch_bootstrap.hpp"
-#endif
 #include "mori/application/bootstrap/socket_bootstrap.hpp"
