+++ include/mori/application/context/context.hpp
@@ -21,6 +21,7 @@
 // SOFTWARE.
 #pragma once
 
+#include <memory>
 #include <string>
 #include <vector>
