+++ include/mori/application/topology/gpu.hpp
@@ -21,6 +21,7 @@
 // SOFTWARE.
 #pragma once
 
+#include <cstdint>
 #include <memory>
 #include <vector>
