+++ include/mori/application/topology/pci.hpp
@@ -25,6 +25,7 @@
 #include <memory>
 #include <string>
 #include <unordered_map>
+#include <utility>
 #include <vector>
 
 #include "mori/application/topology/node.hpp"
