+++ include/mori/core/profiler/kernel_profiler.hpp
@@ -22,7 +22,10 @@
 
 #pragma once
 
+#include <cstdint>  // int64_t / uint8_t
 #include <hip/hip_runtime.h>
+
+#include "mori/core/profiler/constants.hpp"  // MAX_TRACE_EVENTS_PER_WARP, ...
 
 #ifdef ENABLE_PROFILER
 #define IF_ENABLE_PROFILER(x) x
