+++ include/mori/core/transport/rdma/device_primitives.hpp
@@ -21,7 +21,6 @@
 // SOFTWARE.
 #pragma once
 
-// #include "mori/application/transport/rdma/rdma.hpp"
 #include "mori/core/transport/rdma/core_device_types.hpp"
 
 // using namespace mori::application;
