+++ include/mori/core/transport/rdma/providers/bnxt/bnxt_device_primitives.hpp
@@ -23,18 +23,52 @@
 
 #include <hip/hip_runtime.h>
 
+#include <cassert>
+#include <cstddef>
+
 #include "mori/core/transport/p2p/device_primitives.hpp"
 #include "mori/core/transport/rdma/device_primitives.hpp"
 #include "mori/core/transport/rdma/providers/bnxt/bnxt_defs.hpp"
 #include "mori/core/transport/rdma/providers/utils.h"
 #include "mori/core/utils.hpp"
 extern "C" {
-#include "mori/core/transport/rdma/providers/bnxt/bnxt_re_dv.h"
 #include "mori/core/transport/rdma/providers/bnxt/bnxt_re_hsi.h"
 }
 
 namespace mori {
 namespace core {
+
+// BNXT request status -> ibv_wc_status (uses bnxt_re_hsi.h, included above).
+static __device__ __host__ enum ibv_wc_status BnxtHandleErrorCqe(int status) {
+  switch (status) {
+    case BNXT_RE_REQ_ST_OK:
+      return IBV_WC_SUCCESS;
+    case BNXT_RE_REQ_ST_BAD_RESP:
+      return IBV_WC_BAD_RESP_ERR;
+    case BNXT_RE_REQ_ST_LOC_LEN:
+      return IBV_WC_LOC_LEN_ERR;
+    case BNXT_RE_REQ_ST_LOC_QP_OP:
+      return IBV_WC_LOC_QP_OP_ERR;
+    case BNXT_RE_REQ_ST_PROT:
+      return IBV_WC_LOC_PROT_ERR;
+    case BNXT_RE_REQ_ST_MEM_OP:
+      return IBV_WC_LOC_ACCESS_ERR;
+    case BNXT_RE_REQ_ST_REM_INVAL:
+      return IBV_WC_REM_INV_REQ_ERR;
+    case BNXT_RE_REQ_ST_REM_ACC:
+      return IBV_WC_REM_ACCESS_ERR;
+    case BNXT_RE_REQ_ST_REM_OP:
+      return IBV_WC_REM_OP_ERR;
+    case BNXT_RE_REQ_ST_RNR_NAK_XCED:
+      return IBV_WC_RNR_RETRY_EXC_ERR;
+    case BNXT_RE_REQ_ST_TRNSP_XCED:
+      return IBV_WC_RETRY_EXC_ERR;
+    case BNXT_RE_REQ_ST_WR_FLUSH:
+      return IBV_WC_WR_FLUSH_ERR;
+    default:
+      return IBV_WC_GENERAL_ERR;
+  }
+}
 
 /* ---------------------------------------------------------------------------------------------- */
 /*                                         DB Header                                              */
