+++ include/mori/core/transport/rdma/rdma.hpp
@@ -21,14 +21,11 @@
 // SOFTWARE.
 #pragma once
 
-#include "mori/core/transport/rdma/host_primitives.hpp"
+// Host + device RDMA: the host primitives layer plus (on device) the full
+// device primitive + provider stack via rdma_device.hpp. Device-only kernel TUs
+// can include rdma_device.hpp directly to skip the host layer.
 #include "mori/core/transport/rdma/primitives.hpp"
 
 #if defined(__HIPCC__) || defined(__CUDACC__)
-#include "mori/core/transport/rdma/device_primitives.hpp"
-#include "mori/core/transport/rdma/providers/bnxt/bnxt_device_primitives.hpp"
-#include "mori/core/transport/rdma/providers/ionic/ionic_device_primitives.hpp"
-#include "mori/core/transport/rdma/providers/mlx5/mlx5_device_primitives.hpp"
+#include "mori/core/transport/rdma/rdma_device.hpp"
 #endif
-
-#include "mori/core/transport/rdma/providers/mlx5/mlx5_host_primitives.hpp"
